requirements.txt
opencv-python
numpy
matplotlib
albumentations
ultralytics
Pillow
tqdm